适配问好插件
