+++ tests/plugins/morning/test_hello.py
@@ -5,20 +5,18 @@ from tests.fake import fake_group_message_event
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.morning",)], indirect=True)
 async def test_hello_enabled(app: App):
     """测试启动问候已开启的情况"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.morning import hello_cmd, plugin_config
+    require("src.plugins.morning")
+    from src.plugins.morning.plugins.hello import hello_cmd, plugin_config
 
     plugin_config.hello_group_id = [10000]
 
     async with app.test_matcher(hello_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/hello"))
 
         ctx.receive_event(bot, event)
@@ -27,18 +25,16 @@ async def test_hello_enabled(app: App):
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.morning",)], indirect=True)
 async def test_hello_not_enabled(app: App):
     """测试启动问候关闭的情况"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.morning import hello_cmd
+    require("src.plugins.morning")
+    from src.plugins.morning.plugins.hello import hello_cmd, plugin_config
 
     async with app.test_matcher(hello_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/hello"))
 
         ctx.receive_event(bot, event)
@@ -47,20 +43,18 @@ async def test_hello_not_enabled(app: App):
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.morning",)], indirect=True)
 async def test_hello_enable(app: App):
     """测试启动问候，在群里启用的情况"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.morning import hello_cmd, plugin_config
+    require("src.plugins.morning")
+    from src.plugins.morning.plugins.hello import hello_cmd, plugin_config
 
     assert plugin_config.hello_group_id == []
 
     async with app.test_matcher(hello_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/hello 1"))
 
         ctx.receive_event(bot, event)
@@ -71,22 +65,20 @@ async def test_hello_enable(app: App):
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.morning",)], indirect=True)
 async def test_hello_disable(app: App):
     """测试启动问候，在群里关闭的情况"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.morning import hello_cmd, plugin_config
+    require("src.plugins.morning")
+    from src.plugins.morning.plugins.hello import hello_cmd, plugin_config
 
     plugin_config.hello_group_id = [10000]
 
     assert plugin_config.hello_group_id == [10000]
 
     async with app.test_matcher(hello_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/hello 0"))
 
         ctx.receive_event(bot, event)
